tools/requirements.txt
# 图片处理工具依赖包
Pillow>=10.0.0

# 背景移除工具
rembg
onnxruntime
